Tests/LibWeb/Text/expected/Editing/execCommand-insertLinebreak.txt

Before: "foobar"
After: "foo<br>bar"
Before: "<p style="white-space: pre">foobar</p>"
After: "<p style="white-space: pre">foo
bar</p>"
Before: "<p style="white-space: pre">foobar</p>"
After: "<p style="white-space: pre">foobar

</p>"
